feat(Phoenix.PubSub.PG2Server): inter-op with Phoenix PubSub V2
Phoenix PubSub V2 has simplified the PubSub adapters significantly, with
features such as fastlaning no longer required. However, these changes are
backwards compatible with Phoenix PubSub V1.

V2 has the following major protocol changes:

1. The pg2 group has been renamed to `MyPubSub.Adapter` instead of `MyPubSub`
2. The messaging format for `forward_to_local` has a different signature.

This commit introduces the following changes to allow interop between v1 and v2.

1. The PG2 adapter will now subscribe to both the `MyPubSub` and
   `MyPubSub.Adapter` process groups.
2. The Pg2 adapter is capable of handling the v2
   `{:forward_to_local, topic, msg, dispatcher}` in addition to the existing v1
   `{:forward_to_local, fastlane, from_pid, topic, msg}` message, with the
   dispatcher option being ignored.
3. When broadcasting or direct broadcasting, the new v2 message format is
   broadcast to the new `MyPubSub.Adapter` process group members, and a v1
   message to the original `MyPubSub` group members.

This does cause additional work for the PG2 adapter, as it has to fetch
the members once for each group, and do a diff on the members to prevent
duplicate messages being sent to v1 users.

This version is intended to be used as an intermediate release during the
migration from pubsub v1 to v2. Once all nodes have been upgraded to v2, this is
no loger required.